requirements.txt
transformers
evaluate
pandas
datasets
pyyaml
addict
tqdm